phase1/requirements.txt
supabase>=2.3.4
PyMuPDF>=1.23.8
requests>=2.31.0
python-dotenv>=1.0.0
